yaml_config/pyproject.toml
[tool.poetry]
name = "ymlc"
version = "0.0.1"
description = "Update Config with Yaml"
authors = ["OhkuboSGMS <ginger.os777146th@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
absl-py = "^0.13.0"
PyYAML = "^5.4.1"
addict = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
